requirements.txt
absl-py
anyio
astunparse
attrs
beautifulsoup4
bs4
cachetools
certifi
charset-normalizer
click
cmd2
contourpy
cycler
exceptiongroup
ffmpeg-python
flatbuffers
flickrapi
fonttools
future
gast
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
googlesearch-python
grpcio
h11
h2
hpack
httpcore
httplib2
httpx
hyperframe
idna
imagehash
imageio
iniconfig
jax
joblib
kaggle
keras
kiwisolver
lazy-loader
libclang
lime
markdown
markupsafe
matplotlib
ml-dtypes
mtcnn
natsort
networkx
norbert
oauthlib
opencv-contrib-python
opencv-python
opencv-python-headless
opt-einsum
packaging
pandas
pillow
plotly
pluggy
protobuf
pyasn1
pyasn1-modules
pyparsing
pyperclip
pytest
python-dateutil
python-dotenv
python-slugify
pytz
pywavelets
pyyaml
requests
requests-oauthlib
requests-toolbelt
rfc3986
robotexclusionrulesparser
rsa
scikit-image
scikit-learn
scipy
sniffio
soupsieve
tenacity
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow-estimator
termcolor
text-unidecode
threadpoolctl
tifffile
tomli
tqdm
typer
typing-extensions
uritemplate
urllib3
wcwidth
werkzeug
wrapt
